requirements.txt
anyio==4.4.0
asgiref==3.8.1
bidict==0.23.1
boto3==1.35.11
botocore==1.35.11
click==8.1.7
Django==5.1
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
h11==0.14.0
httptools==0.6.1
idna==3.7
jmespath==1.0.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.1
s3transfer==0.10.2
setuptools==74.1.1
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
urllib3==2.2.2
uvicorn==0.30.4
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wheel==0.44.0
wsproto==1.2.0
